+++ sparv/modules/stanza/stanza_swe.py
@@ -37,10 +37,20 @@ def annotate_swe(
         max_token_length: int = Config("stanza.max_token_length")):
     """Do dependency parsing using Stanza."""
     import stanza
+    import torch
     from stanza.pipeline.core import DownloadMethod
 
     # cpu_fallback only makes sense if use_gpu is True
     cpu_fallback = cpu_fallback and use_gpu
+
+    # Select the GPU with most free memory available
+    try:
+        if use_gpu:
+            gpus = util.system.gpus()
+            if gpus:
+                torch.cuda.set_device(gpus[0])
+    except:
+        pass
 
     sentences_all, orphans = sentence.get_children(token)
     if orphans: