+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Added utility function `util.system.gpus()` which returns a list of GPUs, ordered by free memory in descending order.
 - Sparv will automatically order the GPUs in the environment variable `CUDA_VISIBLE_DEVICES` by the amount of free
   memory that was available when Sparv started.
+- Stanza now always selects the GPU with the most free memory.
 
 ### Changed